#0: Move ttnn Resnet to model demos folder
