+++ models/demos/ttnn_resnet/tt/ttnn_functional_resnet50.py
