+++ tests/nightly/wh_b0_only_eth/models/demos/ttnn_resnet/test_ttnn_resnet50_performant.py
@@ -1,0 +1,1 @@
+../../../../../../models/demos/ttnn_resnet/tests/test_ttnn_resnet50_performant.py
