--- tests/nightly/wh_b0_only_eth/models/experimental/resnet/tests/test_ttnn_resnet50_performant.py
@@ -1,1 +1,0 @@
-../../../../../../../models/experimental/resnet/tests/test_ttnn_resnet50_performant.py
