+++ tests/ttnn/integration_tests/resnet/test_ttnn_functional_resnet50_large_new.py
@@ -23,7 +23,7 @@ from models.utility_functions import (
     skip_for_wormhole_b0,
 )
 
-from models.experimental.resnet.tt.ttnn_functional_resnet50_large_new_conv_api import resnet50
+from models.demos.ttnn_resnet.tt.ttnn_functional_resnet50_large_new_conv_api import resnet50
 
 
 def preprocess_conv_parameter(parameter, *, dtype):
