+++ tests/ttnn/integration_tests/resnet/test_ttnn_functional_resnet50_new.py
@@ -22,8 +22,8 @@ from models.utility_functions import (
     skip_for_wormhole_b0,
 )
 
-from models.experimental.resnet.tt.custom_preprocessing import custom_preprocessor
-from models.experimental.resnet.tt.ttnn_functional_resnet50_new_conv_api import resnet50
+from models.demos.ttnn_resnet.tt.custom_preprocessing import custom_preprocessor
+from models.demos.ttnn_resnet.tt.ttnn_functional_resnet50_new_conv_api import resnet50
 
 
 ## copied from ttlib version test:
