+++ tests/ttnn/integration_tests/resnet/test_ttnn_functional_resnet50_xxlarge_new.py
@@ -22,7 +22,7 @@ from models.utility_functions import (
     skip_for_grayskull,
 )
 
-from models.experimental.resnet.tt.ttnn_functional_resnet50_xxlarge_new_conv_api import resnet50
+from models.demos.ttnn_resnet.tt.ttnn_functional_resnet50_xxlarge_new_conv_api import resnet50
 
 
 def preprocess_conv_parameter(parameter, *, dtype):
